fixed the syntax parameter local expand eval issue, heading fixes
The heading grammar needs wsnn+ not wsnn* in almost all cases, there
are some weird edge cases in the elisp implementation that are
confusing, especially with respect to COMMENT.

The heading tokenizer also had to be tweaked, critically to use
from/stop-before for parsing the stars. There is nasty behavior
in the elisp implementation which is due to the fact that it uses
something equivalent to from/to for parsing the stars and as a result
causes confusing behavior when there is an empty title, forcing tags
to become the title, this is fixed in this grammar, and will have to
be revisited in the elisp impelemntation in the future.

h-title expander now uses the paragraph parser, but this has not been
tested yet.

All tests are passing, but there are a number of known issues that the
tests do not currently handle correctly.