first pass fix for file-position misalignment issue
the issue is that parser-tools uses port-next-location instead of
file-position when set-port-count-lines! has been called on a port
we were missing this because that is not done by default but is done
when running as #lang org, dotest-file has been updated accordingly

the alignment issue is likely to remain for other cases and we
probably need to switch from using file-position to using
set-port-next-location! instead, for now the tests pass so ignoring,
but hitting utf-8 chars is likely to introduce a world of pain