major overhaul of markup parsing, first pass footnotes, tests pass
The way we were matching markup syntax resulted in longest matches
instead of shortest matches which is not what the spec demands the
this is now fixed. The main solution was to switch to wrapping all
paragraphs in a starting newline and an ending newline (easier to
do that in do-paragraph than it is to tinker with the port).

We can now mostly parse footnotes but I'm fairly certain that there
are still correctness issues with the outputs and results.

A number of new tests have been added. I haven't written down the
correct behavior, but at this commit the behavior matches elisp.