add table parser, timestamp parser to own file, lex timestamps
The table parser has been its own file for months, but I forgot to add
it to git (oops!).

Moved timestamps out of the main parser and added lexer version.

All of this is generally gearing up to remove the bulk of the top
level grammar and replace it with a much simpler version.