reduce quadratic issues in table lexer/parser by matching spaces
Turns out that at least part of our performance pitfall in the
expander was due to the fact that we were matching multiple spaces
using wssn* where wsnn was usually A SINGLE SPACE :x. Known issue.

Loading some of the slower files is now possible in practically finite
time, there are still some issues lurking in there, but the process is
now fast enough that profiling it is possible.